add custom certs
