+++ ydb/src/auth_test.rs
@@ -4,7 +4,7 @@ use tracing_test::traced_test;
 
 use crate::{
     credentials::StaticCredentials, pub_traits::Credentials, test_helpers::CONNECTION_STRING,
-    test_integration_helper::{create_custom_ca_client, create_password_client}, Query, Transaction, YdbResult,
+    test_integration_helper::create_password_client, Query, Transaction, YdbResult,
 };
 
 #[test]
@@ -92,24 +92,6 @@ async fn wrong_password_test() {
 #[ignore] // YDB access is necessary
 async fn password_client_test() -> YdbResult<()> {
     let client = create_password_client().await?;
-    let two: i32 = client
-        .table_client() // create table client
-        .retry_transaction(|mut t: Box<dyn Transaction>| async move {
-            let res = t.query(Query::from("SELECT 2")).await?;
-            let field_val: i32 = res.into_only_row()?.remove_field(0)?.try_into()?;
-            Ok(field_val)
-        })
-        .await?;
-
-    assert_eq!(two, 2);
-    Ok(())
-}
-
-#[tokio::test]
-#[traced_test]
-#[ignore] // YDB access is necessary
-async fn custom_ca_test() -> YdbResult<()> {
-    let client = create_custom_ca_client().await?;
     let two: i32 = client
         .table_client() // create table client
         .retry_transaction(|mut t: Box<dyn Transaction>| async move {
